# Enable SSH, NX-API and Native Minions in the same nxos namespace