--- a/salt/utils/nxos.py
+++ b/salt/utils/nxos.py
@@ -1,0 +1,273 @@
+# -*- coding: utf-8 -*-
+'''
+Util functions for the NXOS modules.
+'''
+from __future__ import absolute_import, print_function, unicode_literals
+
+# Import Python std lib
+import json
+import logging
+import os
+import sys
+import httplib
+import socket
+import pwd
+import time
+import datetime
+from stat import *
+from salt.exceptions import CommandExecutionError
+
+# Import salt libs
+import salt.utils.http
+from salt.ext import six
+try:
+    from salt.utils.args import clean_kwargs
+except ImportError:
+    from salt.utils import clean_kwargs
+
+log = logging.getLogger(__name__)
+
+class UHTTPConnection(httplib.HTTPConnection):
+    '''
+    Subclass of Python library HTTPConnection that uses a unix-domain socket.
+    '''
+
+    def __init__(self, path):
+        httplib.HTTPConnection.__init__(self, 'localhost')
+        self.path = path
+
+    def connect(self):
+        sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+        sock.connect(self.path)
+        self.sock = sock
+
+
+class NxapiClient():
+    '''
+    Class representing an NX-API client that connects over http(s) or
+    unix domain socket (UDS).
+    '''
+
+    # Location of unix domain socket for NX-API localhost
+    NXAPI_UDS = '/tmp/nginx_local/nginx_1_be_nxapi.sock'
+    # NXAPI listens for remote connections to "http(s)://<switch IP>/ins"
+    # NXAPI listens for local connections to "http(s)://<UDS>/ins_local"
+    NXAPI_REMOTE_URI_PATH = '/ins'
+    NXAPI_UDS_URI_PATH = '/ins_local'
+    NXAPI_VERSION = '1.0'
+
+    def __init__(self, **nxos_kwargs):
+        '''
+        Initialize NxapiClient() connection object.  By default this connects
+        to the local unix domain socket (UDS).  If http(s) is required to
+        connect to a remote device then
+            nxos_kwargs['host'],
+            nxos_kwargs['username'],
+            nxos_kwargs['password'],
+            nxos_kwargs['transport'],
+            nxos_kwargs['port'],
+        parameters must be provided.
+        '''
+        self.nxargs = self._prepare_conn_args(clean_kwargs(**nxos_kwargs))
+        log.info('nxapi connection arguments: {0}'.format(self.nxargs))
+        # Default: Connect to unix domain socket on localhost.
+        if self.nxargs['connect_over_uds']:
+            if not os.path.exists(self.NXAPI_UDS):
+                raise RuntimeError("No host specified and no UDS found at {0}\n".format(self.NXAPI_UDS))
+
+            # Create UHTTPConnection object for NX-API communication over UDS.
+            log.info('Connecting over unix domain socket')
+            self.connection = UHTTPConnection(self.NXAPI_UDS)
+        else:
+            # Remote connection - Proxy Minion, connect over http(s)
+            log.info('Connecting over {}'.format(self.nxargs['transport']))
+            self.connection = salt.utils.http.query
+
+
+    def _use_remote_connection(self, kwargs):
+        '''
+        Determine if connection is local or remote
+        '''
+        kwargs['host'] = kwargs.get('host')
+        kwargs['username'] = kwargs.get('username')
+        kwargs['password'] = kwargs.get('password')
+        if kwargs['host'] is None or \
+           kwargs['username'] is None or \
+           kwargs['password'] is None:
+            return False
+        else:
+            return True
+
+
+    def _prepare_conn_args(self, kwargs):
+        '''
+        Set connection arguments for remote or local connection.
+        '''
+        kwargs['connect_over_uds'] = True
+        kwargs['timeout'] = kwargs.get('timeout', 60)
+        kwargs['cookie'] = kwargs.get('cookie', 'admin')
+        if self._use_remote_connection(kwargs):
+            kwargs['transport'] = kwargs.get('transport', 'https')
+            if kwargs['transport'] == 'https':
+                kwargs['port'] = kwargs.get('port', 443)
+            else:
+                kwargs['port'] = kwargs.get('port', 80)
+            kwargs['verify'] = kwargs.get('verify', True)
+            if isinstance(kwargs['verify'], bool):
+                kwargs['verify_ssl'] = kwargs['verify']
+            else:
+                kwargs['ca_bundle'] = kwargs['verify']
+            kwargs['connect_over_uds'] = False
+        return kwargs
+
+
+    def _build_request(self, type, commands):
+        '''
+        Build NX-API JSON request.
+        '''
+        request = {}
+        headers = {
+            'content-type': 'application/json',
+        }
+        if self.nxargs['connect_over_uds']:
+            user = self.nxargs['cookie']
+            headers['cookie'] = 'nxapi_auth=' + user + ':local'
+            request['url'] = self.NXAPI_UDS_URI_PATH
+        else:
+            request['url'] = '{transport}://{host}:{port}{uri}'.format(
+                transport = self.nxargs['transport'],
+                host = self.nxargs['host'],
+                port = self.nxargs['port'],
+                uri = self.NXAPI_REMOTE_URI_PATH,
+            )
+
+        if isinstance(commands, (list, set, tuple)):
+            commands = ' ; '.join(commands)
+        payload = {}
+        payload['ins_api'] = {
+            'version': self.NXAPI_VERSION,
+            'type': type,
+            'chunk': '0',
+            'sid': '1',
+            'input': commands,
+            'output_format': 'json',
+        }
+        request['headers'] = headers
+        request['payload'] = json.dumps(payload)
+        request['opts'] = {
+            'http_request_timeout': self.nxargs['timeout']
+        }
+        log.info('request: {0}'.format(request))
+        return request
+
+
+    def request(self, type, command_list):
+        '''
+        Send NX-API JSON request to the NX-OS device.
+        '''
+        req = self._build_request(type, command_list)
+        if self.nxargs['connect_over_uds']:
+            self.connection.request('POST', req['url'], req['payload'], req['headers'])
+            response = self.connection.getresponse()
+        else:
+            response = self.connection(req['url'],
+                                        method = 'POST',
+                                        opts = req['opts'],
+                                        data = req['payload'],
+                                        header_dict = req['headers'],
+                                        decode = True,
+                                        decode_type = 'json',
+                                        **self.nxargs)
+            response = response['dict']
+
+        return self.parse_response(response)
+
+
+    def parse_response(self, response):
+        '''
+        Parse NX-API JSON response from the NX-OS device.
+        '''
+        body = response
+        if self.nxargs['connect_over_uds']:
+            body = json.loads(response.read().decode('utf-8'))
+
+        # Proceed with caution.  The JSON may not be complete.
+        # Don't just return body['ins_api']['outputs']['output'] directly.
+        output = body.get('ins_api')
+        if output is None:
+            raise RuntimeError('Unexpected JSON output\n{0}'.format(body))
+        if output.get('outputs'):
+            output = output['outputs']
+        if output.get('output'):
+            output = output['output']
+
+        result = list()
+        if isinstance(output, list):
+            for each in output:
+                log.info('PARSE_RESPONSE: {0} {1}'.format(each.get('code'), each.get('msg')))
+                if each.get('code') != '200':
+                    msg = '{0} {1}'.format(each.get('code'), each.get('msg'))
+                    raise CommandExecutionError(msg)
+                else:
+                    result.append(each['body'])
+        else:
+            log.info('PARSE_RESPONSE: {0} {1}'.format(output.get('code'), output.get('msg')))
+            if output.get('code') != '200':
+                msg = '{0} {1}'.format(output.get('code'), output.get('msg'))
+                raise CommandExecutionError(msg)
+            else:
+                result.append(output['body'])
+
+        return result
+
+
+def nxapi_request(commands,
+        method='cli_show',
+        **kwargs):
+    '''
+    Send exec and config commands to the NX-OS device over NX-API.
+
+    commands
+        The exec or config commands to be sent.
+
+    method:
+        ``cli_show_ascii``: Return raw test or unstructured output.
+        ``cli_show``: Return structured output.
+        ``cli_conf``: Send configuration commands to the device.
+        Defaults to ``cli_show``.
+
+    transport: ``https``
+        Specifies the type of connection transport to use. Valid values for the
+        connection are ``http``, and  ``https``.
+
+    host: ``localhost``
+        The IP address or DNS host name of the device.
+
+    username: ``admin``
+        The username to pass to the device to authenticate the NX-API connection.
+
+    password
+        The password to pass to the device to authenticate the NX-API connection.
+
+    port
+        The TCP port of the endpoint for the NX-API connection. If this keyword is
+        not specified, the default value is automatically determined by the
+        transport type (``80`` for ``http``, or ``443`` for ``https``).
+
+    timeout: ``60``
+        Time in seconds to wait for the device to respond. Default: 60 seconds.
+
+    verify: ``True``
+        Either a boolean, in which case it controls whether we verify the NX-API
+        TLS certificate, or a string, in which case it must be a path to a CA bundle
+        to use. Defaults to ``True``.
+    '''
+    client = NxapiClient(**kwargs)
+    return client.request(method, commands)
+
+
+def ping(**kwargs):
+    '''
+    Verify connection to the NX-OS device over UDS.
+    '''
+    return NxapiClient(**kwargs).nxargs['connect_over_uds']