### What does this PR do?

* Combines `nxos.py` and `nxos_api.py` workflows into the single `nxos` namespace.
* Enables all of the functions defined in `nxos.py`and  `nxos_api.py` to work across the following:
  * NXOS SSH Proxy Minion
  * NXOS NX-API HTTP/HTTPS Proxy Minion
  * NXOS Guestshell Hosted Native Minion over Unix Domain Socket (UDS)
* Performance improvement with `no_save_config` option that disables saving the configuration after each config command is executed.  Provides a new function `save_running_config` that can be used after configuration commands are complete.
* All device config functions now call the `config` function that was ported from `nxos_api.py` that displays the diffset after applying config commands and additionally now displays the command list that was applied to the NX-OS device.
* Provides a new proxy minion parameter called `connection` that is used to control connection over `SSH` or `NX-API`

### Previous Behavior
Previously most of the functions defined in the `nxos.py` SSH proxy minion were not available for the NX-API proxy minion defined in `nxos_api.py`.

### New Behavior
  This PR enables all of the functions defined in both to work across both types of Proxy minions (SSH and NX-API) and additionally provides Native minion support using NX-OS UDS.

### Tests written?

No

### Commits signed with GPG?

Yes

### Tests Executed To Validate The Functionality

The proxy minion pillar file was defined as follows:

This combines options that are needed for both SSH and NX-API proxy minions.

```yaml
proxy:
  proxytype: nxos
  #connection: ssh                <---- Uncomment to enable proxy minion over SSH (Default is SSH)
  #connection: nxapi              <---- Uncomment to enable proxy minion over NX-API
  host: n9k.example.com
  username: admin 
  password: password 
  prompt_name: n9k
  ssh_args: '-o PubkeyAuthentication=no'
  key_accept: True
  transport: http
  port: 80
  verify: False
  no_save_config: True
```

No proxy parameters are needed now for Native Minions hosted inside the GuestShell running on the NX-OS device since it connects over a local Unix Domain Socket.  The GuestShell is only supported on N3k and N9k NX-OS devices running software version `7.0(3)I2(5)` or later.

#### Test File Used to Verify Functionality

```bash
#!/bin/bash

echo_and_run() { echo "$@" ; "$@" ; }

echo_and_run salt $1 nxos.cmd ping

echo_and_run salt $1 nxos.cmd sendline 'show ver'
echo_and_run salt $1 nxos.cmd sendline 'show ver' method='cli_show'
echo_and_run salt $1 nxos.cmd sendline 'show ver' method='cli_show_ascii'

echo_and_run salt $1 nxos.cmd grains
echo_and_run salt $1 nxos.cmd grains_refresh

echo_and_run salt $1 nxos.cmd get_user admin

echo_and_run salt $1 nxos.cmd get_roles admin

echo_and_run salt $1 nxos.cmd check_password admin password='$5$cSAw89/6$fVroF1TQSFK3MogH0dTPHkXIom8KnmXdZeW6ONO4OS1' encrypted=True

echo_and_run salt $1 nxos.cmd check_role admin role=network-admin

echo_and_run salt $1 nxos.cmd remove_user username=devops
echo_and_run salt $1 nxos.cmd add_config 'username devops password cisco123'

echo_and_run salt $1 nxos.cmd set_password devops test123

echo_and_run salt $1 nxos.cmd set_role username=devops role=vdc-admin
echo_and_run salt $1 nxos.cmd get_roles devops
echo_and_run salt $1 nxos.cmd unset_role username=devops role=vdc-admin
echo_and_run salt $1 nxos.cmd get_roles devops

echo_and_run salt $1 nxos.cmd remove_user username=devops
echo_and_run salt $1 nxos.cmd remove_user username=devops


echo_and_run salt $1 nxos.cmd show_run
echo_and_run salt $1 nxos.cmd show_ver

echo_and_run salt $1 nxos.cmd add_config 'feature bgp'
echo_and_run salt $1 nxos.cmd add_config "['feature bgp', 'router bgp 55']"
echo_and_run salt $1 nxos.cmd find '.*bgp.*$'
echo_and_run salt $1 nxos.cmd replace 'feature bgp' 'feature ospf'
echo_and_run salt $1 nxos.cmd find '^feature ospf.*$'
echo_and_run salt $1 nxos.cmd find '^feature bgp.*$'
echo_and_run salt $1 nxos.cmd delete_config 'feature ospf'
echo_and_run salt $1 nxos.cmd find '^feature ospf.*$'

echo_and_run salt $1 nxos.system_info

echo_and_run salt $1 nxos.cmd show 'show inventory'
echo_and_run salt $1 nxos.cmd show 'show inventory' raw_text=False

echo_and_run salt $1 nxos.cmd config 'feature bgp'
echo_and_run salt $1 nxos.config commands="['feature bgp', 'router bgp 55']"
echo_and_run salt $1 nxos.cmd find '.*bgp.*$'
echo_and_run salt $1 nxos.config config_file=salt://bgp_config.txt
echo_and_run salt $1 nxos.config config_file=salt://bgp_config.txt
echo_and_run salt $1 nxos.cmd config 'no feature bgp'
echo_and_run salt $1 nxos.cmd find '^feature bgp.*$'

echo_and_run salt $1 nxos.save_running_config
```

#### Test Results

```
salt n9k-gs nxos.cmd ping
n9k-gs:
    True
salt n9k-gs nxos.cmd sendline show ver
n9k-gs:
    - Cisco Nexus Operating System (NX-OS) Software
      TAC support: http://www.cisco.com/tac
      Copyright (C) 2002-2018, Cisco and/or its affiliates.
      All rights reserved.
      The copyrights to certain works contained in this software are
      owned by other third parties and used and distributed under their own
      licenses, such as open source.  This software is provided "as is," and unless
      otherwise stated, there is no warranty, express or implied, including but not
      limited to warranties of merchantability and fitness for a particular purpose.
      Certain components of this software are licensed under
      the GNU General Public License (GPL) version 2.0 or 
      GNU General Public License (GPL) version 3.0  or the GNU
      Lesser General Public License (LGPL) Version 2.1 or 
      Lesser General Public License (LGPL) Version 2.0. 
      A copy of each such license is available at
      http://www.opensource.org/licenses/gpl-2.0.php and
      http://opensource.org/licenses/gpl-3.0.html and
      http://www.opensource.org/licenses/lgpl-2.1.php and
      http://www.gnu.org/licenses/old-licenses/library.txt.
      
      Software
        BIOS: version 08.34
        NXOS: version 9.2(1)
        BIOS compile time:  04/26/2018
        NXOS image file is: bootflash:///nxos.9.2.1.bin
        NXOS compile time:  7/17/2018 16:00:00 [07/18/2018 00:21:19]
      
      
      Hardware
        cisco Nexus9000 C9504 (4 Slot) Chassis ("Supervisor Module")
        Intel(R) Xeon(R) CPU E5-2403 0 @ 1.80GHz with 16400084 kB of memory.
        Processor Board ID SAL1909A7VC
      
        Device name: n9k
        bootflash:   53298520 kB
      Kernel uptime is 27 day(s), 16 hour(s), 29 minute(s), 22 second(s)
      
      Last reset at 240127 usecs after Tue Aug  7 21:07:10 2018
        Reason: Reset due to upgrade
        System version: 7.0(3)I7(4)
        Service: 
      
      plugin
        Core Plugin, Ethernet Plugin
      
      Active Package(s):
              
salt n9k-gs nxos.cmd sendline show ver method=cli_show
n9k-gs:
    |_
      ----------
      TABLE_package_list:
          ----------
          ROW_package_list:
              ----------
              package_id:
                  ----------
      bios_cmpl_time:
          04/26/2018
      bios_ver_str:
          08.34
      bootflash_size:
          53298520
      chassis_id:
          Nexus9000 C9504 (4 Slot) Chassis
      cpu_name:
          Intel(R) Xeon(R) CPU E5-2403 0 @ 1.80GHz
      header_str:
          Cisco Nexus Operating System (NX-OS) Software
          TAC support: http://www.cisco.com/tac
          Copyright (C) 2002-2018, Cisco and/or its affiliates.
          All rights reserved.
          The copyrights to certain works contained in this software are
          owned by other third parties and used and distributed under their own
          licenses, such as open source.  This software is provided "as is," and unless
          otherwise stated, there is no warranty, express or implied, including but not
          limited to warranties of merchantability and fitness for a particular purpose.
          Certain components of this software are licensed under
          the GNU General Public License (GPL) version 2.0 or 
          GNU General Public License (GPL) version 3.0  or the GNU
          Lesser General Public License (LGPL) Version 2.1 or 
          Lesser General Public License (LGPL) Version 2.0. 
          A copy of each such license is available at
          http://www.opensource.org/licenses/gpl-2.0.php and
          http://opensource.org/licenses/gpl-3.0.html and
          http://www.opensource.org/licenses/lgpl-2.1.php and
          http://www.gnu.org/licenses/old-licenses/library.txt.
      host_name:
          n9k
      kern_uptm_days:
          27
      kern_uptm_hrs:
          16
      kern_uptm_mins:
          29
      kern_uptm_secs:
          24
      kick_cmpl_time:
          7/17/2018 16:00:00
      kick_file_name:
          bootflash:///nxos.9.2.1.bin
      kick_tmstmp:
          07/18/2018 00:21:19
      kickstart_ver_str:
          9.2(1)
      manufacturer:
          Cisco Systems, Inc.
      mem_type:
          kB
      memory:
          16400084
      module_id:
          Supervisor Module
      proc_board_id:
          SAL1909A7VC
      rr_ctime:
          Tue Aug  7 21:07:10 2018
      rr_reason:
          Reset due to upgrade
      rr_service:
      rr_sys_ver:
          7.0(3)I7(4)
      rr_usecs:
          240127
salt n9k-gs nxos.cmd sendline show ver method=cli_show_ascii
n9k-gs:
    - Cisco Nexus Operating System (NX-OS) Software
      TAC support: http://www.cisco.com/tac
      Copyright (C) 2002-2018, Cisco and/or its affiliates.
      All rights reserved.
      The copyrights to certain works contained in this software are
      owned by other third parties and used and distributed under their own
      licenses, such as open source.  This software is provided "as is," and unless
      otherwise stated, there is no warranty, express or implied, including but not
      limited to warranties of merchantability and fitness for a particular purpose.
      Certain components of this software are licensed under
      the GNU General Public License (GPL) version 2.0 or 
      GNU General Public License (GPL) version 3.0  or the GNU
      Lesser General Public License (LGPL) Version 2.1 or 
      Lesser General Public License (LGPL) Version 2.0. 
      A copy of each such license is available at
      http://www.opensource.org/licenses/gpl-2.0.php and
      http://opensource.org/licenses/gpl-3.0.html and
      http://www.opensource.org/licenses/lgpl-2.1.php and
      http://www.gnu.org/licenses/old-licenses/library.txt.
      
      Software
        BIOS: version 08.34
        NXOS: version 9.2(1)
        BIOS compile time:  04/26/2018
        NXOS image file is: bootflash:///nxos.9.2.1.bin
        NXOS compile time:  7/17/2018 16:00:00 [07/18/2018 00:21:19]
      
      
      Hardware
        cisco Nexus9000 C9504 (4 Slot) Chassis ("Supervisor Module")
        Intel(R) Xeon(R) CPU E5-2403 0 @ 1.80GHz with 16400084 kB of memory.
        Processor Board ID SAL1909A7VC
      
        Device name: n9k
        bootflash:   53298520 kB
      Kernel uptime is 27 day(s), 16 hour(s), 29 minute(s), 27 second(s)
      
      Last reset at 240127 usecs after Tue Aug  7 21:07:10 2018
        Reason: Reset due to upgrade
        System version: 7.0(3)I7(4)
        Service: 
      
      plugin
        Core Plugin, Ethernet Plugin
      
      Active Package(s):
              
salt n9k-gs nxos.cmd grains
n9k-gs:
    ----------
    nxos:
        ----------
        hardware:
            ----------
            Device name:
                n9k
            bootflash:
                53298520 kB
        plugins:
            - Core Plugin
            - Ethernet Plugin
        software:
            ----------
            BIOS:
                version 08.34
            BIOS compile time:
                04/26/2018
            NXOS:
                version 9.2(1)
            NXOS compile time:
                7/17/2018 16:00:00 [07/18/2018 00:21:19]
            NXOS image file is:
                bootflash:///nxos.9.2.1.bin
salt n9k-gs nxos.cmd grains_refresh
n9k-gs:
    ----------
    nxos:
        ----------
        hardware:
            ----------
            Device name:
                n9k
            bootflash:
                53298520 kB
        plugins:
            - Core Plugin
            - Ethernet Plugin
        software:
            ----------
            BIOS:
                version 08.34
            BIOS compile time:
                04/26/2018
            NXOS:
                version 9.2(1)
            NXOS compile time:
                7/17/2018 16:00:00 [07/18/2018 00:21:19]
            NXOS image file is:
                bootflash:///nxos.9.2.1.bin
salt n9k-gs nxos.cmd get_user admin
n9k-gs:
    username admin password 5 $5$cSAw89/6$fVroF1TQSFK3MogH0dTPHkXIom8KnmXdZeW6ONO4OS1  role network-admin
salt n9k-gs nxos.cmd get_roles admin
n9k-gs:
    - network-admin
    - priv-15
salt n9k-gs nxos.cmd check_password admin password=$5$cSAw89/6$fVroF1TQSFK3MogH0dTPHkXIom8KnmXdZeW6ONO4OS1 encrypted=True
n9k-gs:
    True
salt n9k-gs nxos.cmd check_role admin role=network-admin
n9k-gs:
    True
salt n9k-gs nxos.cmd remove_user username=devops
n9k-gs:
    COMMAND_LIST: no username devops
    CommandExecutionError('400 CLI execution error',)
salt n9k-gs nxos.cmd add_config username devops password cisco123
n9k-gs:
    COMMAND_LIST: username devops password cisco123
    WARNING: Password should contain characters from at least three of the following classes: lower case letters, upper case letters, digits and special characters.
    WARNING: Configuration accepted because password strength check is disabled
    
    --- 
    +++ 
    @@ -27,6 +27,7 @@
     username admin role priv-15
     username mike password 5 $5$QyvvwC.F$0Cd9wzidODezJp2pN32ek9J5FmEaogYlCmJpHFh8Ts9  role network-admin
    +username devops password 5 $5$VxKg4mY4$XLtTTEZKiCgCjW8bRvF0FLJZIvT1H/134mgLGWn0gs5  role network-operator
     ip domain-lookup
     ip name-server 192.168.1.1 192.168.2.2
     ip access-list my_ipv4_acl
    @@ -46,6 +47,7 @@
     snmp-server user mike network-admin auth md5 0x524f62c6bb01045744039df009274ada priv 0x524f62c6bb01045744039df009274ada localizedkey
     snmp-server user admin network-admin auth md5 0xc128db49c6ca55ae38dbddea78f6299f priv 0xc128db49c6ca55ae38dbddea78f6299f localizedkey
     snmp-server user admin priv-15
    +snmp-server user devops network-operator auth md5 0x60b4ae5dcebe87c4fe850fe558bfd330 priv 0x60b4ae5dcebe87c4fe850fe558bfd330 localizedkey
     rmon event 1 description FATAL(1) owner PMON@FATAL
     rmon event 2 description CRITICAL(2) owner PMON@CRITICAL
salt n9k-gs nxos.cmd set_password devops test123
n9k-gs:
    COMMAND_LIST: username devops password 5 $5$F7W5kWUI$HZQBRms.HgO7bOghGxhfCI2VDXNlokJMIBIKysM./N8
    
    --- 
    +++ 
    @@ -27,7 +27,7 @@
     username admin role priv-15
     username mike password 5 $5$QyvvwC.F$0Cd9wzidODezJp2pN32ek9J5FmEaogYlCmJpHFh8Ts9  role network-admin
    -username devops password 5 $5$VxKg4mY4$XLtTTEZKiCgCjW8bRvF0FLJZIvT1H/134mgLGWn0gs5  role network-operator
    +username devops password 5 $5$F7W5kWUI$HZQBRms.HgO7bOghGxhfCI2VDXNlokJMIBIKysM./N8  role network-operator
     ip domain-lookup
     ip name-server 192.168.1.1 192.168.2.2
     ip access-list my_ipv4_acl
salt n9k-gs nxos.cmd set_role username=devops role=vdc-admin
n9k-gs:
    COMMAND_LIST: username devops role vdc-admin
    
    --- 
    +++ 
    @@ -28,6 +28,7 @@
     username mike password 5 $5$QyvvwC.F$0Cd9wzidODezJp2pN32ek9J5FmEaogYlCmJpHFh8Ts9  role network-admin
     username devops password 5 $5$F7W5kWUI$HZQBRms.HgO7bOghGxhfCI2VDXNlokJMIBIKysM./N8  role network-operator
    +username devops role vdc-admin
     ip domain-lookup
     ip name-server 192.168.1.1 192.168.2.2
     ip access-list my_ipv4_acl
    @@ -48,6 +49,7 @@
     snmp-server user admin network-admin auth md5 0xc128db49c6ca55ae38dbddea78f6299f priv 0xc128db49c6ca55ae38dbddea78f6299f localizedkey
     snmp-server user admin priv-15
     snmp-server user devops network-operator auth md5 0x60b4ae5dcebe87c4fe850fe558bfd330 priv 0x60b4ae5dcebe87c4fe850fe558bfd330 localizedkey
    +snmp-server user devops vdc-admin
     rmon event 1 description FATAL(1) owner PMON@FATAL
     rmon event 2 description CRITICAL(2) owner PMON@CRITICAL
salt n9k-gs nxos.cmd get_roles devops
n9k-gs:
    - network-operator
    - vdc-admin
salt n9k-gs nxos.cmd unset_role username=devops role=vdc-admin
n9k-gs:
    COMMAND_LIST: no username devops role vdc-admin
    
    --- 
    +++ 
    @@ -28,7 +28,6 @@
     username mike password 5 $5$QyvvwC.F$0Cd9wzidODezJp2pN32ek9J5FmEaogYlCmJpHFh8Ts9  role network-admin
     username devops password 5 $5$F7W5kWUI$HZQBRms.HgO7bOghGxhfCI2VDXNlokJMIBIKysM./N8  role network-operator
    -username devops role vdc-admin
     ip domain-lookup
     ip name-server 192.168.1.1 192.168.2.2
     ip access-list my_ipv4_acl
    @@ -49,7 +48,6 @@
     snmp-server user admin network-admin auth md5 0xc128db49c6ca55ae38dbddea78f6299f priv 0xc128db49c6ca55ae38dbddea78f6299f localizedkey
     snmp-server user admin priv-15
     snmp-server user devops network-operator auth md5 0x60b4ae5dcebe87c4fe850fe558bfd330 priv 0x60b4ae5dcebe87c4fe850fe558bfd330 localizedkey
    -snmp-server user devops vdc-admin
     rmon event 1 description FATAL(1) owner PMON@FATAL
     rmon event 2 description CRITICAL(2) owner PMON@CRITICAL
salt n9k-gs nxos.cmd get_roles devops
n9k-gs:
    - network-operator
salt n9k-gs nxos.cmd remove_user username=devops
n9k-gs:
    COMMAND_LIST: no username devops
    
    --- 
    +++ 
    @@ -27,7 +27,6 @@
     username admin role priv-15
     username mike password 5 $5$QyvvwC.F$0Cd9wzidODezJp2pN32ek9J5FmEaogYlCmJpHFh8Ts9  role network-admin
    -username devops password 5 $5$F7W5kWUI$HZQBRms.HgO7bOghGxhfCI2VDXNlokJMIBIKysM./N8  role network-operator
     ip domain-lookup
     ip name-server 192.168.1.1 192.168.2.2
     ip access-list my_ipv4_acl
    @@ -47,7 +46,6 @@
     snmp-server user mike network-admin auth md5 0x524f62c6bb01045744039df009274ada priv 0x524f62c6bb01045744039df009274ada localizedkey
     snmp-server user admin network-admin auth md5 0xc128db49c6ca55ae38dbddea78f6299f priv 0xc128db49c6ca55ae38dbddea78f6299f localizedkey
     snmp-server user admin priv-15
    -snmp-server user devops network-operator auth md5 0x60b4ae5dcebe87c4fe850fe558bfd330 priv 0x60b4ae5dcebe87c4fe850fe558bfd330 localizedkey
     rmon event 1 description FATAL(1) owner PMON@FATAL
     rmon event 2 description CRITICAL(2) owner PMON@CRITICAL
salt n9k-gs nxos.cmd remove_user username=devops
n9k-gs:
    COMMAND_LIST: no username devops
    CommandExecutionError('400 CLI execution error',)
salt n9k-gs nxos.cmd show_run
n9k-gs:
    
    !Command: show running-config
    !Running configuration last done at: Tue Sep  4 13:38:47 2018
    !Time: Tue Sep  4 13:38:50 2018
    
    version 9.2(1) Bios:version 08.34 
    switchname n9k
    vdc n9k id 1
      limit-resource vlan minimum 16 maximum 4094
      limit-resource vrf minimum 2 maximum 4096
      limit-resource port-channel minimum 0 maximum 511
      limit-resource u4route-mem minimum 248 maximum 248
      limit-resource u6route-mem minimum 96 maximum 96
      limit-resource m4route-mem minimum 58 maximum 58
      limit-resource m6route-mem minimum 8 maximum 8
    

<snip... running config>

    interface loopback0
    line console
    line vty
      logout-warning 0
…