@cachedout 2018-09-04
Hi @mikewiebe. Thanks very much for this.

There are a number of lint issues which need to be addressed before we can review this. They are here:

https://jenkinsci.saltstack.com/job/pr-lint/job/PR-49496/1/warnings52Result/new/

Could you please fix them and push your changes? Thanks!